# Use chunked vector to hold mutation vectors