--- a/cql3/query_processor.hh
+++ b/cql3/query_processor.hh
@@ -381,7 +381,7 @@ public:
     // function enables putting multiple CQL queries into a single raft command
     // and vice versa, split mutations from one query into separate commands.
     // It supports write-only queries, read-modified-writes not supported.
-    future<std::vector<mutation>> get_mutations_internal(
+    future<mutation_vector> get_mutations_internal(
         const sstring query_string,
         service::query_state& query_state,
         api::timestamp_type timestamp,
@@ -446,7 +446,7 @@ public:
 
     future<std::string>
     execute_thrift_schema_command(
-            std::function<future<std::vector<mutation>>(data_dictionary::database, api::timestamp_type)> prepare_schema_mutations,
+            std::function<future<mutation_vector>(data_dictionary::database, api::timestamp_type)> prepare_schema_mutations,
             std::string_view description);
 
     std::unique_ptr<statements::prepared_statement> get_statement(