… cql3/statements/batch_statement.cc
… cql3/statements/batch_statement.hh
… cql3/statements/create_aggregate_statement.cc
… cql3/statements/create_aggregate_statement.hh
… cql3/statements/create_function_statement.cc
… cql3/statements/create_function_statement.hh
… cql3/statements/create_index_statement.cc
… cql3/statements/create_index_statement.hh
… cql3/statements/create_keyspace_statement.cc
… cql3/statements/create_keyspace_statement.hh
… cql3/statements/create_table_statement.cc
… cql3/statements/create_table_statement.hh
… cql3/statements/create_type_statement.cc
… cql3/statements/create_type_statement.hh
… cql3/statements/create_view_statement.cc
… cql3/statements/create_view_statement.hh
… cql3/statements/drop_aggregate_statement.cc
… cql3/statements/drop_aggregate_statement.hh
… cql3/statements/drop_function_statement.cc
… cql3/statements/drop_function_statement.hh
… cql3/statements/drop_index_statement.cc
… cql3/statements/drop_index_statement.hh
… cql3/statements/drop_keyspace_statement.cc
… cql3/statements/drop_keyspace_statement.hh
… cql3/statements/drop_table_statement.cc
… cql3/statements/drop_table_statement.hh
… cql3/statements/drop_type_statement.cc
… cql3/statements/drop_type_statement.hh
… cql3/statements/drop_view_statement.cc
… cql3/statements/drop_view_statement.hh