This series converts the use of std::vector holding mutation, frozen_mutation, or canonical_mutation to use utils::chunked_vector instead so to accommodate for long vectors of mutations arriving from the control plane via raft commands for schema, or topology, and in particular for tablets where in some cases we end up with 10s or 100s of thousands of per-tablet mutations.

perf-simple-query results show insignificant improvement in instructions/op.
```
before:median 98014.58 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   42338 insns/op,        0 errors)
 after:median 98323.81 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   42281 insns/op,        0 errors)
```

- [x] ** Backport reason (please explain below if this patch should be backported or not) **
No backport required as this is an improvement required for raft consistent topology changes and tablets that both make their debut in 6.0