@avikivity 2024-05-27
I don't understand why we have long vectors of mutations.