@bhalevy 2025-02-16
Closing this PR as, generally speaking, we want to prevent the long vectors rather than handling them more efficiently.